+++ core/src/models/upload_multipart.rs
@@ -51,7 +51,7 @@ impl CreateFormMultipart {
         self.chunk_processing
             .as_ref()
             .map(|mp_json| mp_json.0.clone())
-            .unwrap_or(ChunkProcessing::default())
+            .unwrap_or_default()
     }
 
     fn get_expires_in(&self) -> Option<i32> {
@@ -146,6 +146,7 @@ impl CreateFormMultipart {
             segment_processing: self.get_segment_processing(),
             segmentation_strategy: self.get_segmentation_strategy(),
             target_chunk_length: None,
+            error_handling: None,
         }
     }
 }
@@ -265,6 +266,7 @@ impl UpdateFormMultipart {
                 .map(|e| e.0.clone())
                 .unwrap_or(current_config.segmentation_strategy.clone()),
             target_chunk_length: None,
+            error_handling: None,
         }
     }
 }